Add 19
vector array in C implementation is added from https://github.com/haxpor/vector_c.
LSpritesheet is added in gl/gl_LTexture_spritesheet.c and gl/gl_LTexture_spritesheet.h in which it accepts gl_LTexture when creatintg a new instance of it, and it will take care freeing of such gl_LTexture as well.
We do in this way to not 100% mimic the way OOP works, and we don't want to rewrite all loading / init functions (boilerplate code) again in new struct, thus we accept underlying system that a new struct rely on.

Hopefully doing this way wil limit number of subclass (or level) that could happen. Let's see if this works in the long run.